Move apply_form_action into bulk_form_actions
Relocate apply_form_action and its FormActionResult/status constants from
utils.py into bulk_form_actions, alongside its only remaining caller now that
archive_or_restore_forms is gone. Pure move; no behavior change.
