+++ corehq/apps/data_interfaces/utils.py
@@ -1,4 +1,3 @@
-from dataclasses import dataclass
 from datetime import datetime, timedelta
 from typing import List, Optional
 
@@ -13,7 +12,7 @@ from soil import DownloadBase
 from corehq.apps.casegroups.models import CommCareCaseGroup
 from corehq.apps.domain.models import Domain
 from corehq.apps.domain_migration_flags.api import any_migrations_in_progress
-from corehq.form_processor.models import CommCareCase, XFormInstance
+from corehq.form_processor.models import CommCareCase
 
 
 def add_cases_to_case_group(domain, case_group_id, uploaded_data, progress_tracker):
@@ -62,49 +61,6 @@ def add_cases_to_case_group(domain, case_group_id, uploaded_data, progress_track
         case_group.save()
 
     return response
-
-
-SUCCEEDED = 'succeeded'
-SKIPPED = 'skipped'
-
-
-@dataclass(frozen=True)
-class FormActionResult:
-    """Outcome of a bulk form action for a single requested form id."""
-    form_id: str
-    status: str  # SUCCEEDED | SKIPPED
-    reason: Optional[str] = None  # not_found | unexpected_error
-
-
-def apply_form_action(domain, form_ids, action_fn, validate=None):
-    """Apply ``action_fn`` to each form and yield a ``FormActionResult`` per id.
-
-    :param action_fn: callable taking an ``XFormInstance``
-    :param validate: optional callable taking an ``XFormInstance`` and returning
-        a skip reason (or ``None`` to proceed).
-    """
-    unresolved_ids = set(form_ids)
-    for xform in XFormInstance.objects.iter_forms(form_ids):
-        if xform.domain != domain:
-            # skip forms not belonging to the specified domain
-            continue
-        unresolved_ids.discard(xform.form_id)
-        reason = validate(xform) if validate else None
-        if reason:
-            yield FormActionResult(xform.form_id, SKIPPED, reason)
-            continue
-        try:
-            action_fn(xform)
-        except Exception:
-            notify_exception(None, "Error applying bulk form action", {
-                'domain': domain,
-                'form_id': xform.form_id,
-            })
-            yield FormActionResult(xform.form_id, SKIPPED, 'unexpected_error')
-        else:
-            yield FormActionResult(xform.form_id, SUCCEEDED)
-    for form_id in unresolved_ids:
-        yield FormActionResult(form_id, SKIPPED, 'not_found')
 
 
 def property_references_parent(case_property):
